RAP-1654 Add DisposableManagerV2 interface
